Move functions to lecture_tmu_2019.utils
